requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
docutils==0.16
entrypoints==0.3
flake8==3.7.9
idna==2.9
isort==4.3.21
Kivy==1.11.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyfrost==0.1.2
Pygments==2.5.2
pylint==2.4.4
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
